requirements.txt
Flask>=2.0
Pillow
numpy
opencv-python
imagehash
scikit-image
